Split out stacking into x11 and wayland
