+++ libqtile/backend/base/__init__.py
@@ -1,7 +1,6 @@
 from libqtile.backend.base.core import Core  # noqa: F401
 from libqtile.backend.base.drawer import Drawer  # noqa: F401
 from libqtile.backend.base.float_states import FloatStates  # noqa: F401
-from libqtile.backend.base.stacking import LayerGroup  # noqa: F401
 from libqtile.backend.base.window import (  # noqa: F401
     Internal,
     Static,
