+++ libqtile/backend/wayland/stacking.py
@@ -1,0 +1,6 @@
+from libqtile.backend.base.stacking import _StackingManager
+
+
+class WaylandStackingManager(_StackingManager):
+    def stack_all(self):
+        pass
